% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTrackPack.R
\docType{package}
\name{myTrackPack}
\alias{myTrackPack}
\alias{myTrackPack-package}
\title{myTrackPack: Routines to put people to sleep during a talk}
\description{
The myTrackPack package is a set of routines designed to totally confuse
people while I drone on about reproducible research. If you are reading this you are in trouble
}
