% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_plot.R
\name{track_plot}
\alias{track_plot}
\title{Map satellite data along track}
\usage{
track_plot(track_df, param, title = "mean of satellite data",
  myColor = cmocean$algae)
}
\arguments{
\item{track_df}{- dataframe combining track and satellite data
outout of `track_frame_merge()`}

\item{param}{- character string with the name of the parameter}

\item{title}{- character string with title for map}

\item{myColor}{- valid R color palette,  default cmocean$algae}
}
\value{
- ggplot graphic
}
\description{
\code{track_plot} plots satellite data along the track
}
\details{
Details:
}
\section{A section of the Details}{
NA
}

\section{More details}{
NA
}

\examples{
urlbase <- 'https://upwell.pfeg.noaa.gov/erddap'
dataInfo <- rerddap::info('erdMBsstd8day')
parameter <- 'sst'
xcoord <- c(230, 231)
ycoord <- c(40, 41)
tcoord <- c('2006-01-15', '2006-01-20')
zcoord <- c(0., 0.)
xlen <- 0.5
ylen <- 0.5
extract <- rxtracto(dataInfo, parameter = parameter, xcoord = xcoord,
                    ycoord = ycoord, tcoord= tcoord, zcoord = zcoord,
                    xlen = xlen, ylen = ylen)
merged_df <- track_frame_merge(xcoord, ycoord, tcoord, extract)
track_plot(track_df, "Chla", title = "Mean Chla along track")
\donttest{
# phony code just to show donttest
xpos <- c(230, 231)
ypos <- c(40, 41)
tpos <- c('2006-05-05', '2006-05-06')
extract <- xtracto_3D('erdMBsstd8day', xpos, ypos, tpos = tpos, verbose=TRUE)
}
}
\references{
https://upwell.pfeg.noaa.gov/erddap/rest.html
}
\author{
Roy Mendelssohn <roy.mendelssohn@noaa.gov>
}
